add default packer that all modules can use
