+++ Jump-exec/WMI/wmi.py
@@ -1,53 +1,10 @@
 from havoc import Demon, RegisterCommand, RegisterModule
 import re
 
-class WmiPacker:
-    def __init__(self):
-        self.buffer : bytes = b''
-        self.size   : int   = 0
-
-    def getbuffer(self):
-        return pack("<L", self.size) + self.buffer
-
-    def addstr(self, s):
-        if s is None:
-            s = ''
-        if isinstance(s, str):
-            s = s.encode("utf-8" )
-        fmt = "<L{}s".format(len(s) + 1)
-        self.buffer += pack(fmt, len(s)+1, s)
-        self.size += calcsize(fmt)
-
-    def addWstr(self, s):
-        s = s.encode("utf-16_le")
-        fmt = "<L{}s".format(len(s) + 2)
-        self.buffer += pack(fmt, len(s)+2, s)
-        self.size += calcsize(fmt)
-
-    def addbytes(self, b):
-        fmt = "<L{}s".format(len(b))
-        self.buffer += pack(fmt, len(b), b)
-        self.size += calcsize(fmt)
-
-    def addbool(self, b):
-        fmt = '<I'
-        self.buffer += pack(fmt, 1 if b else 0)
-        self.size += calcsize(fmt)
-
-    def adduint32(self, n):
-        fmt = '<I'
-        self.buffer += pack(fmt, n)
-        self.size += calcsize(fmt)
-
-    def addshort(self, n):
-        fmt = '<h'
-        self.buffer += pack(fmt, n)
-        self.size += calcsize(fmt)
-
 def wmi_eventsub( demonID, *params ):
     TaskID : str    = None
     demon  : Demon  = None
-    packer = WmiPacker()
+    packer = Packer()
     demon  = Demon( demonID )
 
     if demon.ProcessArch == 'x86':
@@ -105,7 +62,7 @@ def wmi_eventsub( demonID, *params ):
 def wmi_proccreate( demonID, *params ):
     TaskID : str    = None
     demon  : Demon  = None
-    packer = WmiPacker()
+    packer = Packer()
     demon  = Demon( demonID )
 
     if demon.ProcessArch == 'x86':
