+++ Packer/packer.py
@@ -1,0 +1,50 @@
+from struct import pack, calcsize
+
+#
+# this is a helper class
+#
+
+class Packer:
+    def __init__(self):
+        self.buffer : bytes = b''
+        self.size   : int   = 0
+
+    def getbuffer(self):
+        return pack("<L", self.size) + self.buffer
+
+    def addstr(self, s):
+        if isinstance(s, str):
+            s = s.encode("utf-8" )
+        fmt = "<L{}s".format(len(s) + 1)
+        self.buffer += pack(fmt, len(s)+1, s)
+        self.size   += calcsize(fmt)
+
+    def addWstr(self, s):
+        s = s.encode("utf-16_le")
+        fmt = "<L{}s".format(len(s) + 2)
+        self.buffer += pack(fmt, len(s)+2, s)
+        self.size   += calcsize(fmt)
+
+    def addbytes(self, b):
+        fmt = "<L{}s".format(len(b))
+        self.buffer += pack(fmt, len(b), b)
+        self.size   += calcsize(fmt)
+
+    def addbool(self, b):
+        fmt = '<I'
+        self.buffer += pack(fmt, 1 if b else 0)
+        self.size   += calcsize(fmt)
+
+    def adduint32(self, n):
+        fmt = '<I'
+        self.buffer += pack(fmt, n)
+        self.size   += calcsize(fmt)
+
+    def addint(self, dint):
+        self.buffer += pack("<i", dint)
+        self.size   += 4
+
+    def addshort(self, n):
+        fmt = '<h'
+        self.buffer += pack(fmt, n)
+        self.size   += calcsize(fmt)
