+++ SamDump/samdump.py
@@ -1,5 +1,4 @@
 from havoc import Demon, RegisterCommand
-from struct import pack, calcsize
 import re
 import time
 
@@ -8,40 +7,10 @@ import time
 def is_full_path(path):
     return re.match(r'^[a-zA-Z]:\\', path) is not None
 
-class SamDumpPacker:
-    def __init__(self):
-        self.buffer : bytes = b''
-        self.size   : int   = 0
-
-    def getbuffer(self):
-        return pack("<L", self.size) + self.buffer
-
-    def addstr(self, s):
-        if isinstance(s, str):
-            s = s.encode("utf-8" )
-        fmt = "<L{}s".format(len(s) + 1)
-        self.buffer += pack(fmt, len(s)+1, s)
-        self.size += calcsize(fmt)
-
-    def addbytes(self, b):
-        fmt = "<L{}s".format(len(b))
-        self.buffer += pack(fmt, len(b), b)
-        self.size += calcsize(fmt)
-
-    def addbool(self, b):
-        fmt = '<I'
-        self.buffer += pack(fmt, 1 if b else 0)
-        self.size += calcsize(fmt)
-
-    def adduint32(self, n):
-        fmt = '<I'
-        self.buffer += pack(fmt, n)
-        self.size += calcsize(fmt)
-
 def samdump(demonID, *params):
     TaskID : str    = None
     demon  : Demon  = None
-    packer = SamDumpPacker()
+    packer = Packer()
 
     num_params = len(params)
     path = ''
